requirements.txt
opentelemetry-api==1.2.0
opentelemetry-sdk==1.2.0
opentelemetry-exporter-otlp==1.2.0
opentelemetry-opentracing-shim==0.21b0
opentracing==2.4.0
